Revert "Implement Killer & History Heuristic with Late Move Reductions."
