+++ Backend/Engine/MoveSearch.cs
@@ -32,8 +32,6 @@ public class MoveSearch
     public int TableCutoffCount;
     private int TotalNodeSearchCount;
 
-    private readonly HistoryTable HistoryTable = new();
-    private readonly KillerMoveTable KillerMoveTable = new();
     private readonly MoveSearchEffortTable SearchEffort = new();
     private readonly PrincipleVariationTable PvTable = new();
 
@@ -318,7 +316,7 @@ public class MoveSearch
 
         // Allocate memory on the stack to be used for our move-list.
         Span<OrderedMoveEntry> moveSpan = stackalloc OrderedMoveEntry[OrderedMoveList.SIZE];
-        OrderedMoveList moveList = new(ref moveSpan, plyFromRoot, KillerMoveTable, HistoryTable);
+        OrderedMoveList moveList = new(ref moveSpan);
         int moveCount = moveList.NormalMoveGeneration(board, transpositionMove);
         
         if (moveCount == 0) {
@@ -366,9 +364,6 @@ public class MoveSearch
             // replace our alpha with our evaluation.
             alpha = evaluation;
             
-            // Update our history table with our alpha-changing move in hopes we can find similar best moves faster.
-            HistoryTable[board.PieceOnly(move.From), board.ColorToMove, move.To] += depth;
-            
             // Our alpha changed, so it is no longer an unchanged alpha entry.
             transpositionTableEntryType = MoveTranspositionTableEntryType.Exact;
             
@@ -402,14 +397,6 @@ public class MoveSearch
             board.UndoMove(ref rv);
 
             if (!HandleEvaluation(evaluation, ref move)) {
-                if (!board.All(oppositeColor)[move.To] && KillerMoveTable[0, plyFromRoot] != move) {
-                    // Given this move isn't a capture move (quiet move), we store it as a killer move (cutoff move) to
-                    // better sort quiet moves like these in the future, allowing us to achieve a cutoff faster. Also
-                    // make sure we are not saving same move in both of our caches.
-                    KillerMoveTable.ReOrder(plyFromRoot);
-                    KillerMoveTable[0, plyFromRoot] = move;
-                }
-
                 // We had a beta cutoff, hence it's a beta cutoff entry.
                 transpositionTableEntryType = MoveTranspositionTableEntryType.BetaCutoff;
                 break;
@@ -463,7 +450,7 @@ public class MoveSearch
 
         // Allocate memory on the stack to be used for our move-list.
         Span<OrderedMoveEntry> moveSpan = stackalloc OrderedMoveEntry[OrderedMoveList.SIZE];
-        OrderedMoveList moveList = new(ref moveSpan, plyFromRoot, KillerMoveTable, HistoryTable);
+        OrderedMoveList moveList = new(ref moveSpan);
         int moveCount = moveList.QSearchMoveGeneration(board, SearchedMove.Default);
         
         // if (moveCount == 0) {
